Save the bookmarked articles in the local database.

+++ app/src/main/java/com/cancer/yaqeen/data/features/home/articles/responses/BookmarkArticleResponse.kt
@@ -10,10 +10,10 @@ data class BookmarkArticleResponse (
 
 data class InsertBookmarksOneResponse (
     @SerializedName("BookmarkId")
-    val bookmarkID: Long?,
+    val bookmarkID: Int?,
 
     @SerializedName("ContentId")
-    val contentID: Long?,
+    val contentID: Int?,
 
     @SerializedName("CreatedDate")
     val createdDate: String?
